First stab at assistant explore and "mine" pages
