+++ logicle/locales/en/common.json
@@ -83,6 +83,7 @@
   "create_assistant": "Create assistant",
   "create_backend": "Create backend",
   "create_connection": "Create connection",
+  "create_new": "Create new",
   "create_tool_field_apikey_label": "Api key",
   "create_tool_field_apikey_placeholder": "Api key of the ChatGptRetrievalPlugin server...",
   "create_tool_field_baseurl_label": "Base url",
@@ -163,6 +164,7 @@
   "metadata-url": "Metadata URL",
   "metadata-xml": "Metadata XML",
   "model": "Model",
+  "my_assistants": "My assistants",
   "most-active-workspaces": "Most active workspaces",
   "my-profile": "My profile",
   "name": "Name",
